requirements.txt
APScheduler==3.6.3
SQLAlchemy==1.3.16
SQLAlchemy-Utils==0.36.3
psycopg2-binary==2.8.5
tenacity==6.1.0
aiohttp==3.6.2
alembic==1.4.2
Flask==1.1.1
newspaper3k==0.2.8
redis==3.4.1
Flask-Caching==1.8.0
python-dateutil==2.7.3
timeago==1.0.14
